jumping fix
Fixed jumping as vell as added some improvements to existing movement code
