--- Assets/Code/JumpController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c8b508633cc31764ea78e0aa9d884a7f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
