--- Assets/Code/Movable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0e3cef724e5ed2f418ad9585f757fbce
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
