+++ Assets/Code/Movement/GroundChecker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 91ef6d84b8beea04e920c09bbfba91c5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
