+++ Assets/Code/Movement/InputHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 561f308dd35e1d84d83e8b0e18e4efb4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
