+++ Assets/Code/Movement/SurfaceSlider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1558e53c561da84fbf73f9a5abb00d5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
