--- Assets/Code/PlayerInput.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ebf4c6814a62f8a409e8d139e56ec39f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
